backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
openai
python-dotenv
requests
pyjwt
passlib[bcrypt]
xmltodict
